FLux modules and flux integration
