+++ upliftme/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
+    "flux": "^3.1.3",
     "font-awesome": "^4.7.0",
     "react": "^16.2.0",
     "react-background-video-player": "^1.0.1",
@@ -24,6 +25,7 @@
     "eslint-plugin-flowtype": "^2.46.1",
     "eslint-plugin-import": "^2.9.0",
     "eslint-plugin-jsx-a11y": "^5.1.1",
-    "eslint-plugin-react": "^7.7.0"
+    "eslint-plugin-react": "^7.7.0",
+    "gulp": "^3.9.1"
   }
 }
