+++ upliftme/src/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
-import registerServiceWorker from './registerServiceWorker';
+import App from './js/components/App';
 
 ReactDOM.render(<App />, document.getElementById('root'));
